Fix csr_AnnotationMatrix
